Moved create and open projects to ProjectsController
