+++ BugTracker/Controllers/HomeController.cs
@@ -28,32 +28,5 @@ namespace BugTracker.Controllers
 
             return View();
         }
-
-        // Get
-        // Home/CreateProject
-        public ActionResult CreateProject()
-        {
-            return PartialView("~/Views/Home/CreateProject.cshtml");
-        }
-
-        //Post
-        // Home/CreateProject
-        [HttpPost]
-        public ActionResult CreateProject(CreateNewProject model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model); //This needs to return the view with the partial view loaded
-            }
-
-            return View();
-        }
-
-        // Get
-        // Home/_OpenProject
-        public ActionResult OpenProject()
-        {
-            return PartialView("~/Views/Home/_OpenProject.cshtml");
-        }
     }
 }
