+++ BugTracker/Models/ProjectModel.cs
@@ -10,11 +10,11 @@ namespace BugTracker.Models
     public class CreateNewProject
     {
         [Required(AllowEmptyStrings = false, ErrorMessage = "This Field is Required")]
-        [Range(1, 16, ErrorMessage = "Project name must be between 1 and 16 characters")]
+        [StringLength(16, ErrorMessage = "Project name must be between 1 and 16 characters")]
         public string ProjectName { get; set; }
 
-        public int ID { get; set; }
-        public virtual string ApplicationUserID { get; set; }
+        //public int ID { get; set; }
+        //public virtual string ApplicationUserID { get; set; }
 
     }
