+++ BugTracker/Views/Projects/Index.cshtml
@@ -1,0 +1,7 @@
+﻿
+@{
+    ViewBag.Title = "Index";
+}
+
+<h2>Index</h2>
+
